backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
requests==2.32.3
python-dotenv==1.0.1

streamlit==1.36.0
humanize==4.10.0
python-dateutil==2.9.0.post0
